moving element methods to their actions
